# docs/swagger.yaml
definitions:
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.RegisterRequest:
    properties:
      password:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      username:
        type: string
    required:
    - password
    - role
    - username
    type: object
  department.DepartmentCreateReq:
    properties:
      name:
        type: string
      organization_id:
        type: integer
    required:
    - name
    - organization_id
    type: object
  department.departmentResp:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      organization_id:
        type: integer
    type: object
  employee.EmployeeRequest:
    properties:
      contact_info:
        type: string
      department_id:
        type: integer
      name:
        type: string
      position:
        type: string
      remained_day_off:
        type: integer
      salary:
        type: number
      supervisor_employee_id:
        type: integer
      user_id:
        type: integer
    required:
    - contact_info
    - name
    - position
    - remained_day_off
    - salary
    - user_id
    type: object
  employee.EmployeeResp:
    properties:
      contact_info:
        type: string
      department_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      position:
        type: string
      remained_day_off:
        type: integer
      salary:
        type: number
      supervisor_employee_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  entity.Role:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - RoleAdmin
    - RoleEmployee
...
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with username and password to receive a JWT token
      parameters:
      - description: Login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful, returns user ID and JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout by invalidating the JWT token from the server-side
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: User Logout
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with username, password, and role
      parameters:
      - description: Registration request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Registration
      tags:
      - Authentication
  /departments:
    get:
      description: Retrieve a list of all departments
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of departments
          schema:
            items:
              $ref: '#/definitions/department.departmentResp'
            type: array
        "500":
          description: Failed to fetch departments
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all departments
      tags:
      - Department
    post:
      consumes:
      - application/json
      description: Create a new department with the provided details
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Department create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/department.DepartmentCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Department created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create department
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new department
      tags:
      - Department
  /employees:
    get:
      description: Retrieve a list of all employees
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of employees
          schema:
            items:
              $ref: '#/definitions/employee.EmployeeResp'
            type: array
        "500":
          description: Failed to fetch employees
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all employees
      tags:
      - Employee
    post:
      consumes:
      - application/json
      description: Create a new employee with the provided details
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Employee create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employee.EmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Employee created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create employee
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new employee
      tags:
      - Employee
  /employees/{id}:
    delete:
      description: Delete an employee by their ID
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Employee deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid employee ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete employee
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an employee
      tags:
      - Employee
    get:
      description: Retrieve an employee by their ID
      parameters:
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Employee details
          schema:
            $ref: '#/definitions/employee.EmployeeResp'
        "400":
          description: Invalid employee ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Employee not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch employee
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get employee by ID
      tags:
      - Employee
    put:
      consumes:
      - application/json
      description: Update an employee's details
      parameters:
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer token (Format: Bearer ''token'')'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Employee update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/employee.EmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Employee updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update employee
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an employee
      tags:
      - Employee
...
swagger: "2.0"